Find and return the substitutions between two parallel fasta files (each with only a single sequence)

This is for a case like this: you have a fasta file for one chromosome in a species, for example Chinook salmon. You also have another fasta file which holds the bases in coho salmon at the homologous positions in Chinook salmon. (In other words, the coho fasta is exactly the same length as the Chinook fasta.) You want to find all the positions where you have a different base in Chinook than you do in coho, and you want to return those all in a nice data frame.

Usage

fasta_dna_diffs(ref, anc, OutputBigTable = FALSE)

Arguments

ref

path to a fasta (can be gzipped) that contains the "reference" sequence.

anc

path to a fasta (can be gzipped) which represents that homologous sequence, which will often be the "ancstral" sequence which has been aligned to the exact positions in the reference sequence.

OutputBigTable

if TRUE, then this will return a list, element "ret" of which is the standard return value, and element "raw_tib" of which is the full tibble with the full sequences.

Details

Note that upper and lower case letters are treated as the same. So, for example ref = A and anc = a will not be counted as a substitution. Output will be always uppercase.

Additionally, if either sequence has an N, that will not be counted as a difference or a substitution. In fact, all bases that are not A,C,G, or T will be discarded and not recorded as a difference. (In other words, we are just going for substitutions here).

Note, the chromosome name (CHROM in the output) will be taken from the reference fasta (ref).

Note, if the sequence name in ref is of the form >chrom:p1-p2, where p1 and p2 are integers, then the positions of all the reported differences will be shifted to start at p1.

It was surprising to me that these operations could be done so quickly and easily in R.

Examples

ref <- system.file(package = "ecaRbioinf", "extdata", "NC_037124.1_fragment_chinook.fna.gz")
anc <- system.file(package = "ecaRbioinf", "extdata", "NC_037124.1_fragment_coho.fna.gz")
fdd <- fasta_dna_diffs(ref, anc)
#> Sequence starting position: 9661479
#> Percentage substitutions: 0.0214500764427991